Update renderer.js
Here, instead of reloading when the lives are over, I have made a restart function which sets all the variables to default values. Also, I have a High score variable, which appears in the middle of the game window, and is updated whenever the score is greater than current highscore.